requirements.txt
amqp==1.4.9
anyjson==0.3.3
bcrypt==2.0.0
billiard==3.3.0.22
bleach==1.4.2
celery==3.1.19
cffi==1.5.0
Django==1.8.5
django-admin-sortable==2.0.12
django-admin-sortable2==0.6.3
django-appconf==1.0.1
django-celery==3.1.17
django-compressor==1.5
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-email-extras==0.3.2
django-extensions==1.5.7
django-filter==0.11.0
django-forms-builder==0.12.2
django-polymorphic==0.9.1
djangorestframework==3.3.0
djangorestframework-expiring-authtoken==0.1.3
future==0.9.0
html5lib==0.9999999
kombu==3.0.33
Markdown==2.6.3
npm==0.1.1
optional-django==0.1.0
psycopg2==2.6.1
py2-py3-django-email-as-username==1.7.1
pycparser==2.14
pycrypto==2.6.1
python-gnupg==0.3.8
python-memcached==1.57
pytz==2015.7
requests==2.9.1
six==1.10.0
sphinx-me==0.3
sqlparse==0.1.18
Unidecode==0.4.19
